test: create BasePreferencesTest to handle the common needed test code between preferences.
