+++ localDatasource/src/test/java/com/amsterdam/localdatasource/dataStore/AuthenticationLocalDataSourceImplTest.kt
@@ -1,29 +1,13 @@
 package com.amsterdam.localdatasource.dataStore
 
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
 import com.amsterdam.domain.utils.SessionType
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.test.runTest
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import java.io.File
 
-class AuthenticationLocalDataSourceImplTest {
-    private val file: File by lazy {
-        File.createTempFile("test", ".preferences_pb").apply { deleteOnExit() }
-    }
-    private val dataStore: DataStore<Preferences> by lazy {
-        PreferenceDataStoreFactory.create { file }
-    }
+class AuthenticationLocalDataSourceImplTest : BasePreferencesTest() {
     private val authenticationLocalDataSourceImpl by lazy {
         AuthenticationLocalDataSourceImpl(dataStore)
-    }
-
-    @BeforeEach
-    fun setup() {
-        file.writeText("")
     }
 
     @Test
